+++ aura_ui/CHANGELOG.md
@@ -1,4 +1,25 @@
+## [1.0.6]
+
+### Updated
+- Updated MaterialStateProperty(deprecated) to WidgetStateProperty .
+
+
+
 ## [1.0.5] 
 
 ### Added
 - Added new widget 'Sparkle Text'.
+
+
+## [1.0.4]
+
+### Added
+- Added new widget `AuraUICarouselTappable` with onTap functionality.
+
+
+## 1.0.3
+
+-Added Buttons in the Catalog
+- AuiIconButton
+- AuiLoadingButton
+- AuiSimpleButton